# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CommandInput {
  /** @minLength 1 */
  command: string;
}

export interface VngManny {
  id: number;
  name: string;
  /** @nullable */
  currentTask: string | null;
  taskProgressPercent: number;
  integrity: number;
  /** @nullable */
  location?: string | null;
}

export interface VngSectorObject {
  /** @nullable */
  id?: string | null;
  type: string;
  /** @nullable */
  name: string | null;
  summary: string;
  resourceTypes?: string[];
}

export type ProbeStateProbeSectorRelative = {
  x: number;
  y: number;
  z: number;
};

export type ProbeStateProbeSector = {
  relative: ProbeStateProbeSectorRelative;
};

export type ProbeStateProbe = {
  id: number;
  name: string;
  status: string;
  fuelPercent: number;
  integrityPercent: number;
  sector: ProbeStateProbeSector;
};

export interface ProbeState {
  probe: ProbeStateProbe;
  mannies: VngManny[];
  sectorObjects: VngSectorObject[];
  recipes: string[];
}