# lib/api-zod/src/generated/types/commandInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CommandInput {
  /** @minLength 1 */
  command: string;
}