# lib/api-zod/src/generated/types/probeState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProbeStateProbe } from './probeStateProbe';
import type { VngManny } from './vngManny';
import type { VngSectorObject } from './vngSectorObject';

export interface ProbeState {
  probe: ProbeStateProbe;
  mannies: VngManny[];
  sectorObjects: VngSectorObject[];
  recipes: string[];
}