# lib/api-zod/src/generated/types/vngManny.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VngManny {
  id: number;
  name: string;
  /** @nullable */
  currentTask: string | null;
  taskProgressPercent: number;
  integrity: number;
  /** @nullable */
  location?: string | null;
}