# lib/api-zod/src/generated/types/vngSectorObject.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VngSectorObject {
  /** @nullable */
  id?: string | null;
  type: string;
  /** @nullable */
  name: string | null;
  summary: string;
  resourceTypes?: string[];
}